pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.0"
description = "Система для автоматического анализа и обработки алертов с использованием GigaChat API"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
# Основные компоненты LangChain и GigaChat
langchain = "^0.1.0"
langchain-core = "^0.1.0"
langchain-gigachat = "^0.0.4"
langchain-community = "^0.0.10"

# Компоненты LangGraph
langgraph = "^0.0.21"
langgraph-checkpoint = "^0.0.1"
langgraph-cli = "^0.0.1"
langgraph-sdk = "^0.0.1"

# Системные и вспомогательные библиотеки
python-dotenv = "^1.0.0"
requests = "^2.31.0"
urllib3 = "^2.0.7"
base64io = "^1.0.3"
typing-extensions = "^4.5.0"

# Форматирование и UI
rich = "^13.6.0"
rapidfuzz = "^3.0.0"

# Обработка и парсинг данных
regex = "^2023.8.8"
json5 = "^0.9.14"

[tool.poetry.group.dev.dependencies]
# Тестирование
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

# Линтеры и форматтеры
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"

# Документация
sphinx = "^7.2.6"
sphinx-rtd-theme = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aiagent = "src.cli:main"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false 